+++ drools-core/src/main/java/org/drools/core/common/EventFactHandle.java
@@ -63,7 +63,7 @@ public class EventFactHandle extends DefaultFactHandle implements Comparable<Eve
      * @param timestamp the timestamp of the occurrence of this event
      * @param duration the duration of this event. May be 0 (zero) in case this is a primitive event.
      */
-    public EventFactHandle(int id,
+    public EventFactHandle(long id,
                            Object object,
                            long recency,
                            long timestamp,
@@ -72,7 +72,7 @@ public class EventFactHandle extends DefaultFactHandle implements Comparable<Eve
         this( id, object, recency, timestamp, duration, wmEntryPoint, false );
     }
 
-    public EventFactHandle(int id,
+    public EventFactHandle(long id,
                            Object object,
                            long recency,
                            long timestamp,
@@ -92,7 +92,7 @@ public class EventFactHandle extends DefaultFactHandle implements Comparable<Eve
         }
     }
 
-    protected EventFactHandle(int id,
+    protected EventFactHandle(long id,
                               int identityHashCode,
                               Object object,
                               long recency,
